+++ src/shared/types.h
@@ -15,27 +15,27 @@ struct requisicao {
 };
 
 struct requisicao_ack {
-  uint16_t seqn; //Número de sequência que está sendo feito o ack
-  uint16_t num_reqs; // Quantidade de requisições
-  uint16_t total_sum; // Valor da soma agregada até o momento
+  uint64_t seqn; //Número de sequência que está sendo feito o ack
+  uint64_t num_reqs; // Quantidade de requisições
+  uint64_t total_sum; // Valor da soma agregada até o momento
 };
 
 typedef struct __packet {
   uint16_t type; // Tipo do pacote (DESC | REQ | DESC_ACK | REQ_ACK )
-  uint16_t seqn; //Número de sequência de uma requisição
+  uint64_t seqn; //Número de sequência de uma requisição
   struct requisicao req;
   struct requisicao_ack ack;
 } packet;
 
 typedef struct _clients {
   std::string address;
-  int last_req;
-  int last_sum;
+  uint64_t last_req;
+  uint64_t last_sum;
 } clients;
 
 typedef struct _aggregate_sum {
   int num_reqs;
-  int total_sum;
+  uint64_t total_sum;
 } aggregate_sum;
 
 typedef struct _thd_args {
@@ -44,11 +44,11 @@ typedef struct _thd_args {
   int *n_clients;
   char *client_sin_address;
   packet *pack_from_client;
-  long int *total_sum;
+  uint64_t *total_sum;
   int *sockfd;
   struct sockaddr_in *client_addr;
   struct sockaddr_in *server_addr;
   socklen_t client_len;
   pthread_mutex_t *lock;
-  int num_reqs;
+  uint64_t num_reqs;
 } thd_args;
